test: update test case for sync balances
